% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.predict.R
\name{beta.predict}
\alias{beta.predict}
\title{predict neuronal maturation with all module exons using beta regression methods}
\usage{
beta.predict(data)
}
\arguments{
\item{data}{processed splicing matrix with 'NA' and extreme value removed}
}
\value{
data.frame with Predicated maturation stage and Predication confidence score
}
\description{
predict neuronal maturation with all module exons using beta regression methods
}

