% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme.replace.R
\name{extreme.replace}
\alias{extreme.replace}
\title{Replace 0 and 1 in splicing matrix with appoximate value}
\usage{
extreme.replace(data)
}
\arguments{
\item{splicing}{matrix}
}
\value{
splicing matrix with 0 and 1 replaced
}
\description{
Replace 0 and 1 in splicing matrix with appoximate value
}

