% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dis.rbp.R
\name{get.dis.rbp}
\alias{get.dis.rbp}
\title{Calculate distance between reference sample and predicated samples with specifc RBP binding exons}
\usage{
get.dis.rbp(lm.obj, devcortex.rbp)
}
\arguments{
\item{beta}{regression object}
}
\value{
vector of distance to reference samples
}
\description{
Calculate distance between reference sample and predicated samples with specifc RBP binding exons
}

