% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dis.R
\name{get.dis}
\alias{get.dis}
\title{Calculate distance between reference sample and predicated samples}
\usage{
get.dis(lm.obj)
}
\arguments{
\item{beta}{regression object}
}
\value{
vector of distance to reference samples
}
\description{
Calculate distance between reference sample and predicated samples
}

