% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.htmldf.R
\name{get.htmldf}
\alias{get.htmldf}
\title{generate html for data frame}
\usage{
get.htmldf(x, Border = 1, innerBorder = 0, classfirstline = "firstline",
  classfirstcolumn = "firstcolumn", classcellinside = "cellinside",
  append = TRUE, align = "center", captionalign = "bottom",
  classtable = "dataframe", digits = 2, row.names = TRUE, ...)
}
\arguments{
\item{dataframe}{object}
}
\description{
generate html for data frame
}

