% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.modulemat.R
\name{get.modulemat}
\alias{get.modulemat}
\title{Output module exons matrix with or without subtraction of reference sample exon median}
\usage{
get.modulemat(data)
}
\arguments{
\item{original}{splicing matrix}
}
\value{
output two files for following visulization
}
\description{
Output module exons matrix with or without subtraction of reference sample exon median
}

