% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.flex.reg.R
\name{knn.flex.reg}
\alias{knn.flex.reg}
\title{KNN predictor}
\usage{
knn.flex.reg(dis, k)
}
\arguments{
\item{dis}{distance vector output by function get.dis}

\item{k}{parameter for neighbour number}
}
\value{
Maturation stage
}
\description{
KNN predictor
}

