% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturation.predict.R
\name{maturation.predict}
\alias{maturation.predict}
\title{Pipeline function for neuronal prediction}
\usage{
maturation.predict(test.data, outfile, plot = T, verbose = verbose, label)
}
\arguments{
\item{test.data}{input splicing matrix}

\item{outfile}{output file for prediction result}

\item{plot}{default is TRUE to plot PCA for all the data}
}
\description{
Pipeline function for neuronal prediction
}

