% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.replace.R
\name{na.replace}
\alias{na.replace}
\title{Replace NA values with mean of reference sample}
\usage{
na.replace(data)
}
\arguments{
\item{module}{exon splicing matrix}
}
\value{
module exon splicing matrix with NA replaced
}
\description{
Replace NA values with mean of reference sample
}

