% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.html.R
\name{output2html}
\alias{output2html}
\title{generate html file}
\usage{
output2html(content, file = "index.html", append = TRUE)
}
\arguments{
\item{prediction}{result matrix using all exons and RBP-specific exons}
}
\description{
generate html file
}

