% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaplot.R
\name{pcaplot}
\alias{pcaplot}
\title{Plot PCA scatter plot using ggplot2}
\usage{
pcaplot(data.plot, test.plot, plot = T, label)
}
\arguments{
\item{plot}{default is TRUE to output the figure}

\item{user.plot}{data with user data and existing data combined}
}
\description{
Plot PCA scatter plot using ggplot2
}

