% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefilter.R
\name{prefilter}
\alias{prefilter}
\title{Filter module exons by ID}
\usage{
prefilter(data)
}
\arguments{
\item{data}{Data splicing matrix}
}
\value{
Module exon splicing matrix
}
\description{
Filter module exons by ID
}

