% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp.beta.predict.R
\name{rbp.beta.predict}
\alias{rbp.beta.predict}
\title{predict neuronal maturation with all module exons using beta regression methods with RBP specif exons}
\usage{
rbp.beta.predict(data, devcortex.rbp)
}
\arguments{
\item{data}{processed splicing matrix with 'NA' and extreme value removed}
}
\value{
data.frame with Predicated maturation stage and Predication confidence score
}
\description{
predict neuronal maturation with all module exons using beta regression methods with RBP specif exons
}

