% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp.predict.R
\name{rbp.predict}
\alias{rbp.predict}
\title{Pipeline function for neuronal prediction with RBP sepcific exons}
\usage{
rbp.predict(data)
}
\arguments{
\item{data}{input splicing matrix with NA and extreme value replaced}
}
\value{
data.frame with Predicated maturation stage and Predication confidence score
}
\description{
Pipeline function for neuronal prediction with RBP sepcific exons
}

