% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submedian.R
\name{submedian}
\alias{submedian}
\title{Get matrix subtracted by second matrix's median}
\usage{
submedian(data, median.data)
}
\arguments{
\item{data}{}

\item{median.data}{}
}
\value{
matrix
}
\description{
Get matrix subtracted by second matrix's median
}

